+++ src/porofluid_pressure_based_elast/4C_porofluid_pressure_based_elast_dyn.hpp
@@ -12,8 +12,8 @@
 
 FOUR_C_NAMESPACE_OPEN
 
-/*! entry point for the solution of poro multiphase problems problems */
-void poromultiphase_dyn(int restart /* do we have to perform a restart?  */
+/*! entry point for the solution of porofluid-elasticity problems */
+void porofluid_elast_dyn(int restart /* do we have to perform a restart?  */
 );
 
 FOUR_C_NAMESPACE_CLOSE
