+++ src/porofluid_pressure_based_elast/4C_porofluid_pressure_based_elast_input.hpp
@@ -17,7 +17,6 @@
 
 FOUR_C_NAMESPACE_OPEN
 
-// forward declaration
 namespace Core::Conditions
 {
   class ConditionDefinition;
